refactor defaults
