+++ src/frontend/src/defaults/links.tsx
@@ -1,0 +1,24 @@
+import { Trans } from '@lingui/macro';
+
+export const footerLinks = [
+  {
+    link: 'https://inventree.org/',
+    label: <Trans>Website</Trans>,
+    key: 'website'
+  },
+  {
+    link: 'https://github.com/invenhost/InvenTree',
+    label: <Trans>GitHub</Trans>,
+    key: 'github'
+  },
+  {
+    link: 'https://demo.inventree.org/',
+    label: <Trans>Demo</Trans>,
+    key: 'demo'
+  }
+];
+export const navTabs = [{ text: <Trans>Home</Trans>, name: 'home' }];
+
+export const docLinks = {
+  app: 'https://docs.inventree.org/en/latest/app/app/'
+};
